requirements.txt
pytesseract==0.3.10
ffmpeg-python==0.2.0
pytube==12.1.0
